Refactored generate html
