+++ apps/mes/app/components/JobOperation/JobOperation.tsx
@@ -35,7 +35,7 @@ import {
   VStack,
   type JSONContent,
 } from "@carbon/react";
-import { generateHTML } from "@carbon/react/Editor.client";
+import { generateHTML } from "@carbon/react/HTML";
 import { Await, useFetcher, useNavigate, useParams } from "@remix-run/react";
 import { Suspense, useEffect, useMemo, useState } from "react";
 import {
